% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_if_not_exists.R
\name{add_column_if_not_exists}
\alias{add_column_if_not_exists}
\title{Add a column to a data frame if it does not exist}
\usage{
add_column_if_not_exists(df, column_name, default_value = NA)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{column_name}{Name of the column to check or add (string).}

\item{default_value}{Value to fill the column with if created. Defaults to NA.}
}
\value{
A data frame with the column ensured to exist.
}
\description{
Checks whether a column is present in a data frame.
If the column does not exist, it is added with a default value.
}
\examples{
df <- tibble::tibble(a = 1:3)
df <- add_column_if_not_exists(df, "b", 0)

}
