% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_field_flag.R
\name{add_field_flag}
\alias{add_field_flag}
\title{Add field related flags to a data frame based on field notes.}
\usage{
add_field_flag(df)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
relevant field note flags.
}
\description{
A function that checks 3 different and separate conditions related to the
field notes and adds the corresponding flags accordingly. The "sonde not
employed" flag is added if the sonde was not employed at the site.
"site visit" flag is added if the last site visit date is the same as the
date of the data. "sv window" flag is added if site visit flag is detected
within a 75 minute window before (15 mins) and after (60 mins). Note that this function will only work on data
that has already been joined to field notes.
}
\examples{
add_range_flags(df = all_data_flagged$`archery-Actual Conductivity`)
add_range_flags(df = all_data_flagged$`boxelder-Temperature`)
}
