% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_field_notes.R
\name{add_field_notes}
\alias{add_field_notes}
\title{Summarize site parameter data from the API and field notes data frames.}
\usage{
add_field_notes(df, notes)
}
\arguments{
\item{df}{A dataframe with the tidied time series data}

\item{notes}{The munged field notes}

\item{site_arg}{A site name.}

\item{parameter_arg}{A parameter name.}

\item{summarize_interval}{At what time interval the user would like the data set to be aggregated and rounded to. Default is 15 minutes}
}
\value{
A dataframe with summary statistics for a given site parameter data frame, plus field notes data
}
\description{
A function that summarizes and joins site parameter data from the API with the field notes data frames.
}
