% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.R
\name{add_flag}
\alias{add_flag}
\title{Underlying function for flagging data.}
\usage{
add_flag(df, condition_arg, description_arg)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}

\item{condition_arg}{A logical statement that is evaluated in the context of
the data frame.}

\item{description_arg}{Flag added to the \code{flag} column.}
}
\value{
An identical data frame with a \code{flag} column that has been updated with the
flag description provided.
}
\description{
This function adds a flag to the \code{flag} column of a given data frame based on
specified conditions for each row. The name of the flag is provided
by the user.
}
