% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_na_flag.R
\name{add_na_flag}
\alias{add_na_flag}
\title{Add NA flags to a data frame based on \code{mean} column.}
\usage{
add_na_flag(df)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
'missing data' flag.
}
\description{
A function designed to append the 'missing data' flag to a row if the \code{mean}
column in that row contains NA values.
}
\examples{
add_range_flags(df = all_data_flagged$`archery-Actual Conductivity`)
add_range_flags(df = all_data_flagged$`boxelder-Temperature`)
}
