% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_repeat_flag.R
\name{add_repeat_flag}
\alias{add_repeat_flag}
\title{Add a flag if the value in the \code{mean} column repeats sequentially.}
\usage{
add_repeat_flag(df)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
'repeated value' flag.

add_range_flags(df = all_data_flagged$\verb{archery-Actual Conductivity})
add_range_flags(df = all_data_flagged$\code{boxelder-Temperature})
}
\description{
A function designed to append the 'repeated value' flag to a row if the value
in the \code{mean} column is equal to the previous or next value in the \code{mean}
column.
}
