% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spec_flag.R
\name{add_spec_flag}
\alias{add_spec_flag}
\title{Add a flag if the value of a parameter exceeds its sensor specification range.}
\usage{
add_spec_flag(
  df,
  spec_table = yaml::read_yaml("data/qaqc/sensor_spec_thresholds.yml")
)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}

\item{spec_table}{file containing In-Situ spec ranges for each sensor}
}
\value{
A data frame with a \code{flag} column that has been updated with the
'outside of sensor specification range' flag.
}
\description{
A function designed to append the 'outside of sensor specification range' flag to
a row if the \code{mean} column in that row exceeds the sensor specification ranges that are
set in the \code{src/qaqc/sensor_spec_thresholds.yml} file. These are instances
where a value exceeds the expected ranges for the Poudre based on the sensor
manufacturer's specifications.
}
\examples{
add_spec_flag(df = all_data_flagged$`archery-Actual Conductivity`)
add_spec_flag(df = all_data_flagged$`boxelder-Temperature`)

}
\seealso{
\code{\link[=flag_all_data]{flag_all_data()}}
}
