% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suspect_flag.R
\name{add_suspect_flag}
\alias{add_suspect_flag}
\title{Add flags to the \code{flag} column of a dataframe based on large swaths of suspect data, or random values within a wide time period of missing data.}
\usage{
add_suspect_flag(df)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the relevant calculated seasonal range flags.
}
\description{
"24hr anomaly flag" is added if more than 50\% of the data points in a 24 hour window are flagged.
"anomaly window" flag is added if the point is included in a 24hr anomaly.
}
\examples{
add_range_flags(df = all_data_flagged$`archery-Actual Conductivity`)
add_range_flags(df = all_data_flagged$`boxelder-Temperature`)
}
