% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_unsubmerged_flag.R
\name{add_unsubmerged_flag}
\alias{add_unsubmerged_flag}
\title{Add a flag if the sonde was not fully submerged by water.}
\usage{
add_unsubmerged_flag(df)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
'sonde unsubmerged' flag.
}
\description{
A function designed to append the 'unsubmerged' flag to a row if the value
in the \code{relative_depth} column is less than or equal to 0.
}
\examples{
add_unsubmerged_flag(df = all_data_flagged$`archery-Actual Conductivity`)
add_unsubmerged_flag(df = all_data_flagged$`boxelder-Temperature`)

}
\seealso{
\code{\link[=flag_all_data]{flag_all_data()}}
}
