% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_extract_chla_data.R
\name{cal_extract_chla_data}
\alias{cal_extract_chla_data}
\title{Extract Chlorophyll-a data from calibration file html markup}
\usage{
cal_extract_chla_data(div)
}
\description{
Extract Chlorophyll-a data from calibration file html markup
}
