% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_extract_conductivity_data.R
\name{cal_extract_conductivity_data}
\alias{cal_extract_conductivity_data}
\title{Extract Conductivity Data from Calibration Markup}
\usage{
cal_extract_conductivity_data(div)
}
\description{
Extract Conductivity Data from Calibration Markup
}
