% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_extract_fdom_data.R
\name{cal_extract_fdom_data}
\alias{cal_extract_fdom_data}
\title{Extract FDOM Data from Calibration Markup}
\usage{
cal_extract_fdom_data(div)
}
\description{
Extract FDOM Data from Calibration Markup
}
