% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_extract_rdo_data.R
\name{cal_extract_rdo_data}
\alias{cal_extract_rdo_data}
\title{Extract RDO Data from Calibration Markup}
\usage{
cal_extract_rdo_data(div)
}
\description{
Extract RDO Data from Calibration Markup
}
