% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_inv_lm.R
\name{cal_inv_lm}
\alias{cal_inv_lm}
\title{Inverse Linear Model Transformation}
\usage{
cal_inv_lm(df, obs_col, lm_coefs_col)
}
\arguments{
\item{df}{Tibble containing sensor data bounded by two calibrations}

\item{obs_col}{Character string specifying the column name containing
observations recorded in HydroVu}

\item{lm_coefs_col}{Character string specifying the column name containing
linear model coefficients}
}
\description{
Converts sensor observations back to raw instrumental values by applying the
inverse of the linear calibration model. This function reverses the
slope/offset transformation applied by the sensor to recover the original raw
measurements, enabling subsequent re-calibration with updated parameters.
}
\seealso{
\code{\link[=cal_wt]{cal_wt()}}

\code{\link[=cal_lin_trans_lm]{cal_lin_trans_lm()}}
}
