% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_datasets.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine historical and incoming data}
\usage{
combine_datasets(incoming_data_list, historical_data_list)
}
\arguments{
\item{incoming_data_list}{A list of dataframes with the incoming data.}

\item{historical_data_list}{A list of dataframes with the historical data.}
}
\value{
A list of dataframes with the combined historical and incoming data.
}
\description{
A function that combines the subset of historically flagged data and creates historical columns.
NOTE: historical columns should be a part of the historically flagged data, but it is not at this point.
}
\examples{
combine_hist_inc_data(incoming_data_list = incoming_data_list, historical_data_list = historical_data_list)
}
