% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_data_binder.R
\name{final_data_binder}
\alias{final_data_binder}
\title{Update Historical Flag List}
\usage{
final_data_binder(new_flagged_data, historical_flagged_data)
}
\arguments{
\item{new_flagged_data}{A list of data frames that have been flagged.}

\item{historical_flagged_data}{A list of data frames that have been flagged.}
}
\value{
A list of data frames that have been updated with new flagged data.
}
\description{
A function that updates the historical flag list with new flagged data.
}
