% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_do_noise.R
\name{find_do_noise}
\alias{find_do_noise}
\title{Flag dissolved oxygen (DO) sensor noise and potential sonde burial}
\usage{
find_do_noise(df)
}
\arguments{
\item{df}{A site-parameter dataframe containing water quality measurements. Must include:
\itemize{
\item \code{parameter}: Measurement type (function checks for "DO")
\item \code{mean}: The calculated mean value of measurements
\item \code{back1}: The previous measurement value
\item \code{front1}: The next measurement value
\item \code{flag}: Existing quality flags (will be updated by this function)
}}
}
\value{
A dataframe with the same structure as the input, but with \code{flag} column
updated to include "DO interference" and/or "Possible burial" flags where applicable.
}
\description{
This function identifies two types of issues with DO sensors:
\enumerate{
\item Short-term interference: Flagged when sudden fluctuations occur or DO is
abnormally low.
\item Possible burial: Flagged when persistent interference is detected over a
24-hour period.
This function processes site-parameter dataframes, but only modifies those
containg DO data. For those dataframes without DO data, the function returns
the original dataframe unmodified.
}
}
\examples{
# Examples are temporarily disabled
}
\seealso{
\code{\link[=add_flag]{add_flag()}}
}
