% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_calibration.R
\name{fix_calibration}
\alias{fix_calibration}
\title{Correct data with underlying bad calibrations using In-Situ calibration reports}
\usage{
fix_calibration(df, cal_errors)
}
\arguments{
\item{df}{Site-parameter data frame}

\item{cal_errors}{A data frame containing a list of known instances of poor calibration}
}
\value{
A data frame with modified \code{mean} values corrected for improper calibration. These observations are also identified in the \code{cal_fix} column.
}
\description{
This function uses calibration reports collected by field crews to "correct"
instances of poor calibration. These are infrequent, and this function is not yet complete.
}
