% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_statistics.R
\name{generate_summary_statistics}
\alias{generate_summary_statistics}
\title{Generate summary statistics for a given site parameter data frame.}
\usage{
generate_summary_statistics(site_param_df)
}
\arguments{
\item{site_param_df}{A single site-parameter data frame with a \code{mean} column}
}
\value{
A data frame with summary statistics for a given site parameter data frame.
}
\description{
A function that generates summary statistics for a given site parameter data frame. The generated statistics include:
\itemize{
\item The next observation and previous observation values.
\item The rolling 7-point median of the value.
\item The rolling 7-point mean of the value.
\item The rolling 7-point standard deviation of the values.
\item The slope in relation to the observations ahead and behind.
\item The rolling 7-point slope of the observations.
\item The month and year of each data point.
\item The year-month combination.
}
}
\examples{
generate_summary_statistics(site_param_df = all_data_flagged$`archery-Actual Conductivity`)
generate_summary_statistics(site_param_df = all_data_flagged$`boxelder-Temperature`)
}
