% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_dates.R
\name{get_start_dates}
\alias{get_start_dates}
\title{Get the start dates for the HydroVu API pull from the historically flagged data}
\usage{
get_start_dates(incoming_historically_flagged_data_list)
}
\arguments{
\item{incoming_historically_flagged_data_list}{A list of dataframes containing
historical water quality data that has already been processed through the QAQC
workflow. Each list element should represent a site-parameter combination and
be named with the "site-parameter" naming convention.}
}
\value{
A dataframe containing two columns:
\itemize{
\item site: The monitoring location identifier
\item DT_round: The timestamp (in MST timezone) of the most recent Temperature
reading for each site, which will be used as the start date for new API pulls
}
}
\description{
This function finds the most recent timestamp (max datetime) in the Temperature
parameter data frames for each monitoring site's historically flagged data.
Temperature is used as the reference parameter because it is consistently
tracked by all sondes and therefore provides the most reliable indication of
when data was last collected from each site.

The resulting timestamps serve as starting points for new API data requests,
ensuring continuous data collection without gaps or unnecessary duplication.
}
\examples{
# Examples are temporarily disabled
}
\seealso{
\code{\link[=api_puller]{api_puller()}}

\code{\link[=munge_api_data]{munge_api_data()}}
}
