% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mWater.R
\name{load_mWater}
\alias{load_mWater}
\title{Load and tidy mWater field notes}
\usage{
load_mWater(
  creds = yaml::read_yaml("creds/mWaterCreds.yml"),
  summarize_interval = "15 minutes"
)
}
\arguments{
\item{creds}{A .yml file with necessary credentials for accessing the field notes. Contact Sam Struthers if you need access.}

\item{summarize_interval}{At what time interval the user would like the data set to be aggregated and rounded to. Default is 15 minutes.}
}
\value{
A dataframe with the field notes.
}
\description{
A function that uploads and cleans the field notes submitted to mWater.
}
