% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_network_order.R
\name{load_network_order}
\alias{load_network_order}
\title{Load site order definitions from a YAML file}
\usage{
load_network_order(file_path)
}
\arguments{
\item{file_path}{Path to a YAML or CSV file containing site relationships.}
}
\value{
A named list where each element corresponds to a network/system and contains
a character vector of that network's ordered sites from upstream to downstream.
}
\description{
Reads a YAML or CSV file that defines upstream and downstream relationships
for each site and returns a named list of site order vectors. If the site is in an independent network,
the name of the list should be the name of that site to avoid confusion/duplicate searches.

For each network, this function constructs a vector containing the upstream sites
(if defined), the site itself, and the downstream sites (if defined).
}
\details{
\subsection{Example YAML file format}{

See template in inst/templates/template_site_orders.yaml:
\preformatted{
network:
  clp:
    joei, cbri, chd, pfal, pbr, pman, pbd, bellvue, salyer, udall, riverbend, cottonwood, elc, archery, riverbluffs
  springcreek:
    riverbend, springcreek, cottonwood
  boxcreek:
    elc, boxcreek, archery
  sfm:
    sfm
  mtn_campus:
    mtn_campus
}
}

\subsection{Example CSV format}{

See template in inst/templates/template_site_orders.csv:
\preformatted{
network_name,network
clp,"joei, cbri, chd, pfal, pbr, pman, pbd, bellvue, salyer, udall, riverbend, cottonwood, elc, archery, riverbluffs"
springcreek,"riverbend, springcreek, cottonwood"
boxcreek,"elc, boxcreek, archery"
sfm,sfm
mtn_campus,mtn_campus
}
}

\subsection{Expected Outputs}{

\preformatted{
site_order_list <- load_network_order("site_orders.yaml")
site_order_list$clp
# "joei" "cbri" "chd" "pfal" "pbr" "pman" "pbd" "bellvue" "salyer" "udall" "riverbend" "cottonwood" "elc" "archery" "riverbluffs"

site_order_list$sfm
# "sfm"
}
}
}
\seealso{
\code{\link[=network_check]{network_check()}}
}
