% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_old_field_notes.R
\name{load_old_field_notes}
\alias{load_old_field_notes}
\title{Load and tidy old field notes}
\usage{
load_old_field_notes(filepath, summarize_interval = "15 minutes")
}
\arguments{
\item{filepath}{A file path to the raw field notes.}

\item{summarize_interval}{At what time interval the user would like the data set to be aggregated and rounded to. Default is 15 minutes.}
}
\value{
A dataframe with the field notes.
}
\description{
A function that uploads and cleans the field notes excel file. This function adds datetime
columns to the field notes dataframe and filters out field notes where the sensor
was not handled.
}
