% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_threshold_table.R
\name{make_threshold_table}
\alias{make_threshold_table}
\title{Generate Threshold Table for QAQC}
\usage{
make_threshold_table(df)
}
\arguments{
\item{df}{A data frame of cleaned, verified data}
}
\value{
A data frame with the thresholds for the mean, slope_behind, and
standard deviation of the mean for each site and season.
}
\description{
A function designed to generate a threshold table for QAQC. This table
contains the thresholds for the mean, slope_behind, and standard deviation
of the mean for each site and season.
}
\examples{
make_threshold_table(df = all_data_flagged$`archery-Actual Conductivity`)
}
