% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_api_data.R
\name{move_api_data}
\alias{move_api_data}
\title{Move api to archive folder.}
\usage{
move_api_data(api_dir, archive_dir)
}
\arguments{
\item{api_dir}{File path to the api folder}

\item{archive_dir}{File path to the api archive folder}
}
\value{
Nothing returned.
}
\description{
A function designed to move the api data to the archival api folder
}
