% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_api_data.R
\name{munge_api_data}
\alias{munge_api_data}
\title{Munge API data for QAQC workflow}
\usage{
munge_api_data(api_path, network, summarize_interval = "15 minutes")
}
\arguments{
\item{api_path}{Path where the raw API data lives.}

\item{network}{Options include "csu", "all"}

\item{summarize_interval}{At what time interval the user would like the data set to be aggregated and rounded to. Default is 15 minutes.}
}
\value{
A dataframe with the munged API data.
}
\description{
A function designed to munge the raw API data for the QAQC workflow.
}
