% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_insitu_html_log.R
\name{parse_insitu_html_log}
\alias{parse_insitu_html_log}
\title{Parse In-Situ log HTML File}
\usage{
parse_insitu_html_log(html_markup)
}
\arguments{
\item{html_markup}{HTML markup from an In Situ Log file created by VuSitu.}
}
\value{
A data frame containing the parsed sensor data for the HTML markup.
}
\description{
Parses an Insitu HTML file to extract sensor data with HTML extraction methods.
}
\examples{
\dontrun{
# Example requires HTML file
html_content <- rvest::read_html("path/to/file.html")
result <- parse_insitu_html_log(html_content)
}
}
