% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update master data list}
\usage{
update_data(new_flagged_data, historical_flagged_data)
}
\arguments{
\item{new_flagged_data}{A list of data frames that have been flagged.}

\item{historical_flagged_data}{A list of data frames that have been flagged.}
}
\value{
A list of data frames that have been updated with new flagged data.
}
\description{
A function that updates the historical flag list with new flagged data.
}
