% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_burial_flag.R
\name{add_burial_flag}
\alias{add_burial_flag}
\title{add_burial_flag}
\usage{
add_burial_flag(df)
}
\arguments{
\item{df}{An updated data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
DO-related possible burial flag.
}
\description{
A function designed to identify instances of likely sonde burial
}
\seealso{
\code{\link[=add_flag]{add_flag()}}
}
