% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_depth_shift_flag.R
\name{add_depth_shift_flag}
\alias{add_depth_shift_flag}
\title{Add a depth shift flag}
\usage{
add_depth_shift_flag(df, level_shift_table, post2024 = TRUE)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}

\item{level_shift_table}{The data frame containing information about when the sonde was shifted.}

\item{post2024}{Whether the data in the QAQC pipeline is before (FALSE) or after (TRUE) the 2024 field season.}
}
\value{
A data frame with a \code{flag} column that has been updated with the "sonde moved" flag.
}
\description{
A function for determining when sonde was moved in its housing
}
