% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_repeat_flag.R
\name{add_repeat_flag}
\alias{add_repeat_flag}
\title{Flag sequentially repeated measurements}
\usage{
add_repeat_flag(df)
}
\arguments{
\item{df}{A data frame containing water quality measurements. Must include columns:
\itemize{
\item \code{mean}: The calculated mean value of measurements
\item \code{front1}: The next measurement value
\item \code{back1}: The previous measurement value
\item \code{flag}: Existing quality flags (will be updated by this function)
}}
}
\value{
A data frame with the same structure as the input, but with the \code{flag}
column updated to include "repeated value" for any measurements that match
either the preceding or following value.
}
\description{
Identifies and flags water quality measurements that have the exact same value
as either the preceding or following measurement. Repeated values can indicate
sensor malfunction, signal processing issues, or data transmission problems.
}
\examples{
# Examples are temporarily disabled
}
\seealso{
\code{\link[=add_flag]{add_flag()}}
}
