% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_seasonal_flag.R
\name{add_seasonal_flag}
\alias{add_seasonal_flag}
\title{Add flags related to calculated parameter seasonal ranges.}
\usage{
add_seasonal_flag(df, threshold_table)
}
\arguments{
\item{df}{A data frame with a \code{flag} column.}

\item{threshold_table}{A dataframe containing seasonal threshold values.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
relevant calculated seasonal range flags.
}
\description{
A function that checks 2 different and separate conditions related to the
calculated parameter seasonal ranges. The 'outside of seasonal range' flag is
added if the \code{mean} value is outside the seasonal 1st - 99th percentile. The
'slope violation' flag is added if the \code{slope_ahead} or \code{slope_behind} value
is greater than or equal to the \code{t_slope_behind_99} value, or if the
\code{slope_ahead} or \code{slope_behind} value is less than or equal to the
\code{t_slope_behind_01} value. The \code{t_slope_behind_99} and \code{t_slope_behind_01}
values are specific to each site-parameter.
}
