% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_puller.R
\name{api_puller}
\alias{api_puller}
\title{Download water quality data from HydroVu API}
\usage{
api_puller(
  site,
  start_dt,
  end_dt = Sys.time(),
  api_token,
  hv_sites_arg = hv_sites,
  dump_dir,
  synapse_env = FALSE,
  fs = NULL
)
}
\arguments{
\item{site}{Character string or vector specifying the site name(s) to download
data for. Site names are matched case-insensitively against the HydroVu
location names.}

\item{start_dt}{POSIXct timestamp indicating the starting point for data
retrieval. Usually derived from the most recent timestamp in existing
historical data.}

\item{end_dt}{POSIXct timestamp indicating the endpoint for data retrieval.
Default is the current system time (Sys.time()).}

\item{api_token}{OAuth client object obtained from hv_auth() function, used
for authentication with the HydroVu API.}

\item{dump_dir}{Character string specifying the directory path where
downloaded CSV files should be saved.}

\item{fs}{Logical, whether to use the file system functions}
}
\value{
No direct return value. The function writes CSV files to the specified
dump_dir, with filenames formatted as "sitename_timestamp.csv".
}
\description{
Downloads raw water quality monitoring data from the HydroVu platform for
specified sites and time periods. This function handles the connection to
HydroVu, processes the API responses, and saves the raw data as parquet files
in a specified directory.
}
\examples{
# Examples are temporarily disabled
}
\seealso{
\code{\link[=get_start_dates]{get_start_dates()}}

\code{\link[=hv_auth]{hv_auth()}}

\code{\link[=munge_api_data]{munge_api_data()}}
}
