% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_puller.R
\name{api_puller}
\alias{api_puller}
\title{Download API data from HydroVu}
\usage{
api_puller(site, network, start_dt, end_dt = Sys.time(), api_token, dump_dir)
}
\arguments{
\item{site}{List of sites to download data for.}

\item{start_dt}{Start DT for downloading data.}

\item{end_dt}{End DT for downloading data. Default is the current time.}

\item{api_token}{API token for accessing HydroVu. Pulled in via \code{hv_auth()}.}

\item{dump_dir}{Directory where raw API data will be saved to.}
}
\description{
A function designed to download raw API data from HydroVu and store in a user-selected file location.
}
