% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_do_noise.R
\name{find_do_noise}
\alias{find_do_noise}
\title{Flag dissolved oxygen (DO) sensor noise and potential sonde burial}
\usage{
find_do_noise(df)
}
\arguments{
\item{df}{A site-parameter dataframe containing water quality parameters with
a \code{flag} column. Must include summary statistics columns (\code{back1}, \code{front1},
\code{mean}) generated by \code{generate_summary_statistics()}.}
}
\value{
A dataframe with the same structure as its input, with additional flags
for DO interference and possible burial.
}
\description{
This function identifies two types of issues with DO sensors:
\enumerate{
\item Short-term interference: Flagged when sudden fluctuations occur or DO is
abnormally low.
\item Possible burial: Flagged when persistent interference is detected over a
24-hour period.
The function processes site-parameter dataframes, but only modifies those
containg DO data. For those dataframes without DO data, the function returns
the original dataframe unmodified.
}
}
\examples{
df <- find_do_noise(df = all_data_summary_stats_list$`riverbluffs-DO`)
}
