% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_dates.R
\name{get_start_dates}
\alias{get_start_dates}
\title{Get the start dates for the HydroVu API pull from historically flagged data}
\usage{
get_start_dates(incoming_historically_flagged_data_list)
}
\arguments{
\item{incoming_historically_flagged_data_list}{A list of dataframes containing
historical water quality data that has been processed through the QAQC
workflow. Each list element should represent a site-parameter combination and
be named using the "site-parameter" convention.}
}
\value{
A dataframe with three columns:
\itemize{
\item site: The monitoring location identifier
\item start_DT: The timestamp (in UTC) of the most recent reading for each site
\item end_DT: The current system time (in UTC)
}
}
\description{
Attempts to find the most recent timestamp in Temperature parameter data for each monitoring
site's historically flagged data. Temperature is used as the reference parameter
because it is consistently tracked by all sondes, providing the most reliable
indication of when data was last collected from each site. This function will compare
the temperature data's most recent timestamp to that of all the other parameters.
If it finds that these are not equal, it will prefer to use the earliest timestamp,
regardless of its source, with the assumption that the earlier start dates should be
attributed to parameters that are not Temperature.

The resulting timestamps serve as starting points for new API data requests,
ensuring continuous data collection without gaps or duplication.
}
\examples{
# Examples are temporarily disabled
}
\seealso{
\code{\link[=api_puller]{api_puller()}}

\code{\link[=munge_api_data]{munge_api_data()}}
}
