% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_dates.R
\name{get_start_dates}
\alias{get_start_dates}
\title{Get the start dates for the HydroVu API pull from the historically
flagged data}
\usage{
get_start_dates(incoming_historically_flagged_data_list)
}
\arguments{
\item{incoming_historically_flagged_data_list}{This is the output from
\code{flagged_data_dfs} and is the historically flagged data.}
}
\value{
A data frame that has the sites and their corresponding starting
date-times' for the HydroVu API pull. This data frame will then get passed into
\code{incoming_data_csvs_upload}.
}
\description{
This function finds the max datetime in the temperature data
frames for each site's historically flagged data. We use temperature because
it is always being tracked in the sensors and thus is the most accurate
representation of the last API pull for a given sensor.
}
