% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersensor_check.R
\name{intersensor_check}
\alias{intersensor_check}
\title{intersensor_check}
\usage{
intersensor_check(df)
}
\arguments{
\item{df}{An updated data frame with a \code{flag} column.}
}
\value{
A data frame with a \code{flag} column that has been updated with the
inter-sensor flag reduction step.
}
\description{
A function designed to reduce overflagging: if a slope violation occurs at the
same time as a slope violation in either depth or temperature, it is likely
not a sensor malfunction and instead a real product of the river.
}
\seealso{
\code{\link[=add_flag]{add_flag()}}
}
