% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_check.R
\name{network_check}
\alias{network_check}
\title{Network Check}
\usage{
network_check(df, network = "all")
}
\arguments{
\item{df}{A site-parameter data frame that has gone through the initial flagging
process.}

\item{network}{Whether the network check is happening across all sites or CSU sites.}
}
\value{
A modified data frame flags that have been altered based on the network check.
}
\description{
This function performs a network check on a given data frame, flagging potential
issues in the data based on upstream and downstream sites.
}
\examples{
network_check(df = all_data_flagged$`archery-Actual Conductivity`)
}
