% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_check.R
\name{network_check}
\alias{network_check}
\title{Reduce overflagging by comparing across monitoring network}
\usage{
network_check(
  df,
  intrasensor_flags_arg = intrasensor_flags,
  site_order_arg = site_order_list
)
}
\arguments{
\item{df}{A site-parameter dataframe that has undergone initial flagging. Must include:
\itemize{
\item \code{site}: Standardized site name
\item \code{parameter}: The measurement type
\item \code{DT_round}: Timestamp for measurements
\item \code{flag}: Existing quality flags
}}

\item{intrasensor_flags_arg}{A list of data frames that have gone through the
intra-sensor flagging functions which are indexed by their corresponding
site-parameter combination.}

\item{site_order_arg}{A list defining the order of sites in the network. This requires
a user create a yaml or csv file with formatting consistent
with the example in \code{load_site_order()} and to create a site_order_list object using \code{load_site_order()}}
}
\value{
A dataframe with the same structure as the input, plus an \code{auto_flag} column
that contains cleaned flags where network-wide events have been accounted for.
}
\description{
Identifies and corrects overflagging by comparing data patterns across upstream
and downstream monitoring sites. When quality flags at a site coincide with similar
patterns at neighboring sites, these flags are often indicating real water quality
events rather than sensor malfunctions and can be removed.

Different site configurations are handled based on the monitoring network specified.
For chlorophyll measurements in the FCW/CSU network, no changes are made as these
measurements are site-specific.
}
\examples{
# Examples are temporarily disabled
}
\seealso{
\code{\link[=load_site_order]{load_site_order()}}
}
